Extract header utils
